fuse: Implement multi-conn
If multi-conn is possible, then open multiple connections to the NBD
server.  Each of these is associated with a background operations
thread which is responsible for dispatching commands on that thread.

This is quite a lot faster than simply using parallel commands on a
single connection.

   READ: bw=302MiB/s (316MB/s), 75.4MiB/s-75.5MiB/s (79.1MB/s-79.1MB/s), io=4096MiB (4295MB), run=13570-13582msec

(Compare this to the stats shown in commit 3f257eb60)