fuse: Better handling of filesystem unmount
Previously we had a handler for release() which happens every time a
program closes the file, not when the filesystem is unmounted
(although likely release() would happen just before unmounting).

Instead, hook the destroy() method which is called just before
unmount.

Also don't bother calling nbd_flush on unmount, and don't busy wait in
a tight loop for commands to complete.