fuse: Further fix race condition when kicking operation thread
The previous fix still left a small window during which the start
condition could be missed, resulting a very rare but reproducible
deadlock.  Extend the lock around the whole loop to prevent this.

However even this is not completely sufficient.  It was possible that
operations threads would be slower to start than the FUSE threads, in
which case there could still be a deadlock where the FUSE threads
would signal the start condition before the operation thread had even
got to that point.  The solution is to add a barrier so we don't start
FUSE processing until the operations threads are running.

This whole thing needs to be rethought because the lock section is too
large, resulting in FUSE threads not being able to issue a command
unless there are no other commands in flight.  But at least for the
moment it should no longer deadlock.

Fixes: commit 594d54c87713990e6b0da63da7e9eecdc405ed85
Fixes: commit 3f257eb60636f498649dc41af6a3d6996bd393bd