tests: Only set MALLOC_CHECK_ and MALLOC_PERTURB_ in tests.
Previously these variables were set in the ./run script.  However this
script has a dual purpose, as both a place to run libnbd tools without
installation, and a way to run the tests.  When benchmarking, we don't
want malloc-check being used because of the overhead, but we still
want to run everything from the local directory for fast iteration.

Solution is to move these variables to TESTS_ENVIRONMENT so they only
run from tests.

This also fixes an incidental bug: MALLOC_PERTURB_ would not be set
correctly if awk was not installed, but awk should not be a
requirement of libnbd.  (./configure replaces @AWK@ anyway which is
why it happened to work.)  We do require bash so use the bash $RANDOM
function instead.