info/nbdinfo.c: Don't leak output buffer
It's interesting why valgrind didn't report this as a leak.  That's
because main calls exit(3), so the output buffer which is a local
variable of main is still reachable when the program is exiting.
Valgrind does not default to giving errors for still reachable memory
(see valgrind --errors-for-leak-kinds parameter).  If the program had
called return instead of exit then valgrind would have shown an error.

(cherry picked from commit ed9f07d72661d1953b7f25d20ed6bee2c71e2f57)

For stable-1.6 branch note the original commit summary was:

  info/main.c: Don't leak output buffer