tests: Require cap_sys_admin where root is required
Some tests check for root privileges, but do not check if they have the
appropriate capabilities for all the tasks they are going to perform.  Not being
sure which ones are the proper ones all these currently check for cap_sys_admin,
although the code is made so that it can check for any capability that is
requested.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>