ci: Some build script fixes
Provide a default for `$MAKE` so that the script is more easily usable outside
of CI and propagate `$CONFIGURE_OPTS` from lcitool to the configure script.
Lastly describe why cross compilation checking is skipped.  Even though cross
compilation is completely disabled in CI, the more generic of a script we have
the easier the adaptation to other projects could be.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>