tests: Rename version to get-version
FreeBSD has (and possibly other platforms as well) a header file named "version"
in the system C++ headers, particularly `/usr/include/c++/v1/version`, which is
supposed to get picked up during the compilation of compile-cxx test as it gets
included by other system files.  However, because there is also a "version" file
in the tests/ directory (one of the tests to check nbd_get_version() function)
it prefers that one for some reason.  And the C preprocessor gets thoroughly
confused, mainly by comments in the libtool script that is ./version, so this
commit changes the `version` binary name to `get-version` to avoid the issue.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>