Include sys/vsock.h if linux/vm_sockets.h is not available
This is the case on MacOS for example.  The issue is that AF_VSOCK is defined in
sys/socket.h, but struct sockaddr_vm is not, so the safeguard lower in the file
does not work.  Thankfully sys/vsock.h has the sockaddr_vm definition, so
including that fixes the build issue.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>