info: Avoid ambiguous 'allocated' terminology in --map
Qemu has unfortunately used two distinct meanings for 'allocated' over
time, and that ambiguity has overloaded the word to make it difficult
to know which sense is meant in a given context.  In one usage,
'allocated' implies that data is present and occupies space on a block
device (the opposite of sparse).  In another usage, 'allocated'
implies that guest contents come from the current layer of a qcow2
backing chain (even if via a sparse zero cluster), rather than
deferring to the backing file.  It doesn't help that in qemu source
code, BDRV_BLOCK_ALLOCATED refers solely to this latter usage, while
a preallocated qcow2 image refers to the former usage.

But we can help by using better terminology in nbdinfo output.  With
base:allocation, either a cluster is known to be sparse (a hole), or
it (likely) occupies space (remember, NBD block status bits are
advisory - if they are set to 1, we know a given property is true; but
if they are left 0, we cannot assume the opposite of the property).
Instead of calling the latter allocated, we can call it merely "data".
And this maps a bit better to 'qemu-img map --output=json' output; a
sparse section really is "data":false, and everything that is not
known to be sparse is "data":true.

Similarly, in qemu:allocation-depth, either a cluster is known to come
from the backing chain (either locally, or from a given backing
depth), or it is completely absent from the chain (not allocated in
any of the layers).  Instead of calling it "unallocated", we can call
it "absent".

Adjust the testsuite and to match our improved terminology.
Hopefully, nbdinfo is new enough that there won't be too many
command-line users affected by the change in output.  For the
human-readable output, it's not much of a concern, and even for the
json output, machines should stick to the numeric flag value rather
than parsing the string since we did not document which strings would
be produced other than in examples.

Thanks: Nir Soffer