ci: Rework the build script to run check-valgrind properly
In order to run check-valgrind we need to skip the broken tests twice, so move
that code into its own functions to clean up.  After that run check-valgrind by
default.  It is still skipped on some platforms where it is not working at the
moment and it does not look like it is related to libnbd.

Also regenerate the container and cirrus files as the libnbd patches for
libvirt-ci were finally pushed to master.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>