python: Implement nbd.aio_connect for AF_UNIX
This call previously just called abort().  Implement it for Unix
domain sockets (the easy case).  Implementing it for AF_INET and
AF_INET6 is more complicated so that is left as a to-do.

Note also that implementing this fully for Python is a bit pointless.
It would be easier for a Python program to call
nbd.aio_connect_tcp(host, port) instead of calling
nbd.aio_connect((host, port)).  Both cases would do hostname lookups
but the former is already implemented.