ci: Allow failure when building fedora rawhide container
I never really considered rawhide stable enough for too many CI jobs, although
it is nice to have some bleeding-edge updates checked as soon as possible.
However I did not want to stray away from some thought-through setup that was
done in libvirt when switching to libvirt-ci.  Nevertheless Fedora Rawhide is
now not being able to build and it is not because of our wrongdoing and libvirt
switched the container building phase to allow failures, which I concur and
hence have no reason for keeping it as it was.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>