copy: Fix progress bar
In multi-thread mode we were using the wrong size (the destination,
not the source) to calculate the percent copied.  When copying from a
small disk to a large disk, even though we only copy the length of the
small disk, the percentage would stay at 0%.

Fixes virt-v2v -o null (which uses a max-sized nbdkit null instance as
the destination).