tests: Prefer qemu-nbd --cache=writeback
QEMU 6.1 and earlier default qemu-img to --cache=writeback, but
qemu-nbd to --cache=writethrough, which is inherently slower because
it requires more flushing.  Although qemu 6.2 will probably be
changing the default of qemu-nbd for consistency, in the meantime we
can be explicit in our choice for slightly faster testing.

Tests that only read from the image are not altered, as the slowdowns
really only impact writing to qemu-nbd.  Documentation examples were
not updated, because there we can just wait for new-enough qemu-nbd
with saner defaults.

Based on a qemu-nbd report by Nir Soffer.

(cherry picked from commit 099dfa073deb98401f3165703337a37cdece26e1)