golang: Fix invalid use of unsafe.Pointer
In golang 1.17 we got this problem with callbacks:

  runtime: bad pointer in frame libguestfs.org/libnbd.(*Libnbd).BlockStatus.func1 at 0xc000051d50: 0x6e
  fatal error: invalid pointer found on stack

It seems like our use of unsafe.Pointer to directly convert the
callback ID (an int) into a pointer to store in void *user_data is not
permitted.  (See: https://pkg.go.dev/unsafe#Pointer)

Following similar code in libvirt-go, I changed the generated code to
malloc a block on the C heap to store the callback ID and use this as
the void *user_data pointer.