interop: Set up GnuTLS debugging in tests
It was hard to tell if we were really using TLS from the debug
messages (especially when testing if qemu-nbd + TLS was working).  The
solution is to enable GnuTLS debugging.

Unfortunately GnuTLS debugging is flawed in several ways (see comment
in code) and that prevents us from adding this as a generic API to
libnbd.  This change is sufficient for testing only.