nbdsh: Do not document --connect
This is an old alias for -u/--uri, which was confusingly named.  In
the comments it said it was undocumented, but it still appeared in the
manual, so remove that documentation completely.

(cherry picked from commit 78ed20a5e37cf7c2dd6ad63b1f1a6d727fa3f8d5)