lib: Allow tls-certificates=<DIR> query parameter in URIs
For nbd_connect_uri, this allows a non-default path to a certificates
directory to be specified.  For example:

  nbds+unix://user@/?socket=/tmp/sock&tls-certificates=tests/pki

nbd_get_uri is also extended to produce the tls-certificates query
field if nbd_set_tls_certificates was called.

The main work here is extending the test suite so it actually tests
TLS URIs properly.  Firstly we need to add --tls-verify-peer to the
nbdkit command line so it checks TLS client credentials at all
(previously it enabled TLS but didn't verify the client).  Then we
need to add tests which use TLS certificates (previously only PSK was
being tested).  And finally I loosened the rules for comparing URIs
since the order that query strings are returned by nbd_get_uri is not
necessarily the same as the query strings in nbd_connect_uri.