python: Fix crash when passing None as a buffer
$ nbdkit -U - null 512 --run 'nbdsh -u $uri -c "h.pwrite(None, 0)"'
nbdkit: external command was killed by signal 11

The stack trace indicated that the code was calling
PyBuffer_Release (&buf) on an uninitialized buffer 'buf'.  What was
happening was that PyArg_ParseTuple was recognizing the type error in
the parameter (None is not a buffer-like object) and failing, but it
didn't initialize the Py_buffer parameter so that contained random
stuff from the stack, and then we tried to release it by calling
PyBuffer_Release.

To fix this, initialize the stack buffer, then avoid calling
PyBuffer_Release if the .obj field in the Py_buffer is still NULL
(which should be impossible if PyArg_ParseTuple set the &buf
parameter).

After this commit we see the type error:

$ nbdkit -U - null 512 --run './run nbdsh -u $uri -c "h.pwrite(None, 0)"'
Traceback (most recent call last):
  File "/usr/lib64/python3.10/runpy.py", line 196, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/usr/lib64/python3.10/runpy.py", line 86, in _run_code
    exec(code, run_globals)
  File "/home/rjones/d/libnbd/python/nbd.py", line 2568, in <module>
    nbdsh.shell()
  File "/home/rjones/d/libnbd/python/nbdsh.py", line 138, in shell
    exec(c, d, d)
  File "<string>", line 1, in <module>
  File "/home/rjones/d/libnbd/python/nbd.py", line 1610, in pwrite
    return libnbdmod.pwrite(self._o, buf, offset, flags)
TypeError: a bytes-like object is required, not 'NoneType'

(cherry picked from commit 51195424f8ba73d0ab1fb4145ddfa81cb8056d4e)