golang: Add README.md and LICENSE files
Add README.md and LICENSE file to the module directory. This helps
people consuming this code from the tarball, which seems to be the only
way to consume now.

The LICENSE is unmodified LPGL2.1 text from:
https://www.gnu.org/licenses/old-licenses/lgpl-2.1.txt

It is same as COPYING.lib, without the examples license notice, since
the go module directory do not include examples.

Preview in gitlab:
https://gitlab.com/nirs/libnbd/-/tree/1b5b2e48ce8fadf1acbb3f2cd1afdbe93a867f3c/golang/src/libguestfs.org/libnbd

Signed-off-by: Nir Soffer <nsoffer@redhat.com>