common/utils: Add vector benchmarks
The generic vector reallocs on every append. Add benchmarks to measure
the cost with uint32 vector (used for copying extents) and the effect of
reserving space upfront.

The tests show that realloc is pretty efficient, but calling reserve
before the appends speeds the appends up significantly.

$ ./test-vector | grep bench
bench_reserve: 1000000 appends in 0.004159 s
bench_append: 1000000 appends in 0.014897 s

Signed-off-by: Nir Soffer <nsoffer@redhat.com>