common/utils: Do not run benchmarks by default
The new benchmarks are very quick on modern machine, but in overload CI,
under valgrind, or running in qemu emulation they can be too slow to
run by default.

Run the benchmarks only when LIBNBD_BENCH environment variable is
defined.

To run all benchmarks in the project:

    make bench

To run single test program with benchmarks:

    $ LIBNBD_BENCH=1 ./test-vector
    bench_reserve: 1000000 appends in 0.004282 s
    bench_append: 1000000 appends in 0.015586 s

CI jobs that run on reliable environment (x86_64?) can add the `bench`
target to the job, to ensure that the benchmarks do not rot.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>