configure: golang: Cleanup golang check
Move the configure test module to configure/test.go, and create new
module for this directory, so it does not clash with the real libnbd
module, that I want to move up to the golang/ directory.

Instead of running "go version", print runtime.Version() in the test
program.

    $ go run golang/configure/test.go
    go1.16.8

Instead of running "go help modules" for checking if modules are
supported, using "go mod tidy". This will fail if the installed go
version is to old, since the go.mod file specify go 1.13, like the real
libnbd go.mod.

    $ (cd golang/configure/ && go mod tidy)
    go mod tidy: go.mod file indicates go 1.17, but maximum supported version is 1.16

I did not try to verify the go version, since it seems impossible given
the crazy versioning. However if we want to verify the version we can do
this using the test output outside of go.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>