golang: Flatten source tree
Move the libnbd module to the golang directory to make it easier to work
with the source, and to match other bindings.

I think that src/libguestfs.org/libnbd/ tree is left over from older Go
versions using GOPATH. Modern Go uses modules so it does not matter
where you put your source.

With this change, we can add a nice tag for every release:

    golang/v1.11.2

This exposes the package version to "go get".

This will also fix:

    https://pkg.go.dev/github.com/libguestfs/libnbd/golang

This is not enough to consume the module directly from the source, we
still have to solve the issue of the generated files, and returning the
right metadata from https://libguestfs.org/libnbd.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>