socket-activation: Allow multiple connections
When using qemu-nbd with systemd socket activation, trying to connect
additional clients concurrently fails with:

    nbd_connect_uri: connect: server backlog overflowed, see
    https://bugzilla.redhat.com/1925045: Resource temporarily
    unavailable

But this bug was fixed in qemu-nbd long time ago; turns out that libnbd
has the same bug, creating the server socket with:

    listen(s, 1);

Use SOMAXCONN so we can have multiple connections for better
performance. I think this can be very useful for nbdcopy.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>