common/utils/vector: Better overflow handling
Check newcap * itemsize for overflow prior to calling realloc, so that
we don't accidentally truncate an existing array.  Rewrite the scaling
by 1.5 to be less prone to intermediate overflow.  Set errno to ENOMEM
on all failure paths, rather than leaving it indeterminate on
overflow.  The fix depends on a prerequisite assumption that
v->cap*itemsize has not previously overflowed.

Fixes: 985dfa72ae (common/utils/vector.c: Optimize vector append)