maint: Suggest better diff output for API.ml
Git diff is able to customize the regex used to locate "function
headers", or the text to output on @@ lines of a patch to make it
easier to determine which portion of a file the patch touches.  This
is done by coupling .gitattributes contents with the user running the
right 'git config' command(s).  We don't have a bootstrap script, so
we don't have an automated place to forcefully ensure the right setup
for new contributors, but we can at least document the steps a
developer can do to make our lives easier.

For the API.ml file, not only do we want to highlight typical
top-level OCaml 'let' statements, we also want to highlight the start
of each of our API definitions.

Thanks: Laszlo Ersek <lersek@redhat.com>