tests: Simplify tests involving qcow2 dirty bitmaps
Rely on 'qemu-img bitmap' instead of $QEMU_BINARY to create the dirty
bitmap in the qcow2 images under test.  This limits the test to qemu
5.1 and newer, but in the long run is easier to maintain.

Nothing in the codebase now uses requires_qemu, although I deferred
removal of that function to a separate patch to make it easier to
revert if we want to revive it.