ocaml/examples: Fix extents example
Laszlo Ersek noted that the original example we had did not work
properly if the size of the disk was larger than around 2G.

The nbd_block_status API is really difficult to use correctly!  In
particular it is not guaranteed that the server will return extents
covering the size requested.  It's also not guaranteed that a bad
server will return any base:allocation extents at all (although such a
server would not be conforming - the protocol says that servers must
always make forward progress).

This commit attempts a fix, although it is not complete especially if
the server is badly behaved.  It also makes the output look a bit
better by aligning the columns.  Also we use nbdkit-sparse-random-
plugin with a larger size to test the > 2G case.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>