ocaml: Add NBD.with_handle utility function
  NBD.with_handle (fun nbd -> ...)

calls the inner function with a newly created handle, and ensures that
NBD.close is always called even if the inner function throws an
exception.

This is inspired by Laszlo Ersek's similar function added to virt-v2v
lib/utils.ml in the commit below.  Unlike that, this does not abstract
the connection and does not call NBD.shutdown.

https://github.com/libguestfs/virt-v2v/commit/b4a8ccf00f1364d703c6d5cf1fd77850105fdd65

I adjusted one example and a few of the tests to use this function.
It is not really necessary to use it, so having a mix of both styles
seems reasonable.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>
(cherry picked from commit 5663ee5b6e628647d87ac1833092410266964ce9)