Add .editorconfig file
This project uses non standard 2 spaces indent in some parts (lib, copy,
info) and tabs in others (e.g. golang/*.go). This can be solved by local
configuration or adding metadata comments to files, but there is a
better solution - editorconfig[1].

Add .editorconfig file that use 2 spaces indent by default, and tabs for
the Go sources. This is not complete yet, but good enough for the areas
I usually touch.

People that do not use an editorconfig plugin are not affected by this
change. People that want to use it can install a plugin[2][3] and
improve the configuration as needed.

[1] https://editorconfig.org/
[2] https://github.com/editorconfig/editorconfig-vim#readme
[3] https://github.com/editorconfig/editorconfig-emacs#readme

Signed-off-by: Nir Soffer <nsoffer@redhat.com>