examples: glib-main-loop: don't strand commands when gsource disappears
Returning 0 from a callback handler implies that we will later call
nbd_aio_command_completed(), but we weren't doing that, and our
finalize() handler instead asserts that no outstanding commands
remain.  Even when we are going to ignore a late completion (due to
the gsource already being torn down), we should still auto-retire the
command.

Fixes: 7e53d48e ("api: Allow completion callbacks to auto-retire", v0.1.9)