golang: tests: Fix error handling
Like lot of the C examples, the aio copy test ignores read and write
errors in the completion callback, which can cause silent data
corruption. The failure in the test is not critical, but this is a bad
example that may be copied by developers to a real application.

The test panics now if completion callback fails, similar to other Go
examples.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>
(cherry picked from commit b80780e980275a879c13d27aff1449f91f883ce6)