python: tests: Fix error handling
Like a lot of the C examples, the aio copy test ignores read and write
errors in the completion callback, which can cause silent data
corruption. The failure in the test is not critical, but this is a bad
example that may be copied by developers to a real application.

The test now dies with a RuntimeError if completion callback fails,
also demonstrating how get at the errno value.  Tested with the
temporary patch of:

| diff --git i/python/t/590-aio-copy.py w/python/t/590-aio-copy.py
| index 861fa6c8..4cd64d83 100644
| --- i/python/t/590-aio-copy.py
| +++ w/python/t/590-aio-copy.py
| @@ -117,7 +117,8 @@ src.set_handle_name("src")
|  dst = nbd.NBD()
|  dst.set_handle_name("dst")
|  src.connect_command(["nbdkit", "-s", "--exit-with-parent", "-r",
| -                     "pattern", "size=%d" % disk_size])
| +                     "--filter=error", "pattern", "error-pread-rate=1",
| +                     "size=%d" % disk_size])
|  dst.connect_command(["nbdkit", "-s", "--exit-with-parent",
|                       "memory", "size=%d" % disk_size])
|  asynch_copy(src, dst)

Message-Id: <20220203202558.203013-2-eblake@redhat.com>
Thanks: Nir Soffer <nsoffer@redhat.com>
(cherry picked from commit e0151cbba74612929979dda374ce17d837b665f4)