ocaml: tests: Fix error handling
Like a lot of the C examples, the aio copy test ignores read and write
errors in the completion callback, which can cause silent data
corruption. The failure in the test is not critical, but this is a bad
example that may be copied by developers to a real application.

The test dies with an assertion failure now if completion callback
fails.  Tested with the temporary patch of:

| diff --git i/ocaml/tests/test_590_aio_copy.ml w/ocaml/tests/test_590_aio_copy.ml
| index a0339c8b..332bf31b 100644
| --- i/ocaml/tests/test_590_aio_copy.ml
| +++ w/ocaml/tests/test_590_aio_copy.ml
| @@ -120,7 +120,8 @@ let
|    let dst = NBD.create () in
|    NBD.set_handle_name dst "dst";
|    NBD.connect_command src ["nbdkit"; "-s"; "--exit-with-parent"; "-r";
| -                           "pattern"; sprintf "size=%d" disk_size];
| +                           "--filter=error"; "pattern"; "error-pread-rate=1";
| +                           sprintf "size=%d" disk_size];
|    NBD.connect_command dst ["nbdkit"; "-s"; "--exit-with-parent";
|                             "memory"; sprintf "size=%d" disk_size];
|    asynch_copy src dst

Message-Id: <20220203202558.203013-3-eblake@redhat.com>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>
(cherry picked from commit 2dbd41a75ef5a275ef35a1c46626ab11813a397c)