golang: make-dist.sh: Generate the list file
Generated the list file when creating the distribution. Since the Go
tool treat the list file on the proxy server as the source of truth, we
do the same. The new list file is created by downloading the current
list file, sorting it, and appending the current version.

Creating a distribution tarball requires now access to
download.libguestfs.org.

With this change the distribution tarball can be extract on the server
without any additional manual process.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>