api: Guarantee sanitized buffer on pread failure
As mentioned in the previous patch, we left the state of the buffer
undefined if we fail pread prior to attempting NBD_CMD_READ.  Better
is to tweak the generator to sanitize the buffer unconditionally, as a
way of hardening against potential bugs in client applications that
fail to check for errors, but which should not be leaking
uninitialized data.  In the process, we can document that the contents
of buf are now merely unspecified, rather than undefined (valgrind
will no longer complain if you read buf, regardless of how nbd_pread
failed).

Since always calling memset() can be a performance hit if the client
also sanitizes the buffer or is willing to take the audit risk, the
next patch will add a knob that allows the client to control when this
happens.

This patch causes changes to just the four APIs
nbd_{aio_,}pread{,_structured}, with the generated lib/api.c changes
looking like:

| --- lib/api.c.bak       2022-02-10 08:15:05.418371357 -0600
| +++ lib/api.c   2022-02-10 08:15:13.224372023 -0600
| @@ -2871,6 +2871,7 @@ nbd_pread (struct nbd_handle *h, void *b
|      debug (h, "enter: buf=<buf> count=%zu offset=%" PRIu64 " flags=0x%x", count, offset, flags);
|    }
|
| +  memset (buf, 0, count);
|    if (unlikely (!pread_in_permitted_state (h))) {
|      ret = -1;
|      goto out;

Message-Id: <20220209220726.1902761-3-eblake@redhat.com>
Acked-by: Laszlo Ersek <lersek@redhat.com>
[eblake: enhance commit message to show generated file diff]
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>

(cherry picked from commit c7fbc71de54bf4e477011176197ad22950a02aa3)